refactor(crdt-core): split Document impl into focused submodules
